backend/requirements.txt
django==6.0.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
